requirements.txt
selenium==4.20.0
Pillow==10.2.0
